requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.14.0
webdriver-manager==4.0.1
langchain>=0.1.0
langchain-huggingface>=0.0.1
huggingface_hub>=0.16.4
python-dotenv==1.0.0
streamlit==1.27.2
# tiktoken==0.5.1  # Commented out as it requires Rust compiler
chromadb==0.4.18
PyPDF2==3.0.1
transformers>=4.30.0
sentence-transformers>=2.2.2